+++ CorpMediaJson/src/com/intuiture/corp/json/ProfessionalAllowanceJson.java
@@ -3,7 +3,7 @@ package com.intuiture.corp.json;
 public class ProfessionalAllowanceJson {
 
 	private Integer professionalAllowanceId;
-	private Long maxAnnualLimit;
+	private Double maxAnnualLimit;
 	private Integer requireSubmissionId;
 	private Integer companyId;
 
@@ -15,11 +15,11 @@ public class ProfessionalAllowanceJson {
 		this.professionalAllowanceId = professionalAllowanceId;
 	}
 
-	public Long getMaxAnnualLimit() {
+	public Double getMaxAnnualLimit() {
 		return maxAnnualLimit;
 	}
 
-	public void setMaxAnnualLimit(Long maxAnnualLimit) {
+	public void setMaxAnnualLimit(Double maxAnnualLimit) {
 		this.maxAnnualLimit = maxAnnualLimit;
 	}
